requirements.txt
beautifulsoup4>=4.9.3
google-cloud-storage>=1.40.0
lxml==4.6.3
regex==2021.4.4
requests==2.25.1
tqdm==4.61.1
